requirements.txt
sentence-transformers==3.4.1
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .
pymongo[srv]==3.11
flask-bcrypt
flask-session
google-genai
google-generativeai